Consolidate RedisDict and RedisStorageDict into single implementation
- Remove duplicate RedisDict class from dict.py (was unused)
- Enhance RedisStorageDict with read_all_values() and close() methods
- Update all imports and test references to use RedisStorageDict
- Fix test configuration to remove obsolete Redis patching
- All 132 storage tests passing
- Eliminates schema drift and duplicate logic maintenance
